Add Locate Button - Add Funcionality to Locate Button - Fix Issues PR 26 Co-authored-by: AlexF090 <AlexF090@users.noreply.github.com>
