+++ client/src/pages/MapPage.jsx
@@ -1,34 +1,13 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import stations from '../data/db.js';
 import Header from '../components/Header.jsx';
-import Map from '../components/Map.jsx';
-import { useMapEvents, Marker, Popup } from 'react-leaflet';
+import Map from '../components/map/Map.jsx';
 
 function MapPage({ gasInfoHead, title, fuelValue }) {
-  function LocationMarker() {
-    const [position, setPosition] = useState(null);
-
-    useEffect(() => {
-      map.locate();
-    }, []);
-
-    const map = useMapEvents({
-      locationfound(e) {
-        setPosition(e.latlng);
-        map.flyTo(e.latlng, map.getZoom());
-      },
-    });
-    return position === null ? null : (
-      <Marker position={position}>
-        <Popup> Du bist hier !</Popup>
-      </Marker>
-    );
-  }
-
   return (
     <>
       <Header title={title} gasInfoHead={gasInfoHead} fuelValue={fuelValue} />
-      <Map LocationMarker={LocationMarker} stations={stations} fuelValue={fuelValue} />
+      <Map stations={stations} fuelValue={fuelValue} />
     </>
   );
 }
